pointers arrays strings

+++ 0x05-pointers_arrays_strings/3-puts.c
@@ -10,7 +10,7 @@
 void _puts(char *str)
 {
 	int i;
-	for(i = 0 ; i < strlen(str) ; i++)
+	for(i = 0 ; i < (unsigned int)strlen(str) ; i++)
 	{
 		_putchar(str[i]);
 	}
